quantai/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
redis==5.0.4
xgboost==2.0.3
scikit-learn==1.4.2
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
alembic==1.13.1
pydantic-settings==2.2.1
MetaTrader5==5.0.45
python-telegram-bot==21.3
apscheduler==3.10.4
structlog==24.2.0
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
fakeredis==2.23.0
